feat(deploy): Add liveliness check
This introduces a liveliness check that can be run after a deployment to
ensure the new generation is healthy. If the check fails, the deployment
is automatically rolled back to the previous generation.

The liveliness check is configured via the `liveliness_check_command`
option.

In addition, this commit also includes:

- Updates to Go dependencies.
- A version bump to 0.8.1.
- The addition of `.envrc` and `.gitignore` files.