feat(scheduler): add random delay to poller
This commit introduces a new option 'random_delay' to the poller
configuration. This allows to specify a maximum random delay in
seconds before fetching the remote, which helps to avoid the
'thundering herd' problem when managing a large fleet of devices.

The documentation and NixOS module options have been updated
accordingly.